+++ test/smokes/cpplint/default/foo.cc
@@ -1,0 +1,5 @@
+// Copyright 2019 Your Company
+bool foo()
+{
+	return 1;
+}