+++ test/smokes/cpplint/option_target/src/foo.cpp
@@ -1,0 +1,2 @@
+// Copyright 2019 Foo
+void foo(){}