+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.7.5...HEAD)
 
+- Bugsnag notify on `rescue` block [#445](https://github.com/sider/runners/pull/445)
+
 ## 0.7.5
 
 [Full diff](https://github.com/sider/runners/compare/0.7.4...0.7.5)